Monolítico/requirements.txt
flask
gunicorn
transformers
torch --index-url https://download.pytorch.org/whl/cpu
faiss-cpu
numpy
